refactor(auth): 將 jwt token 拉出放在 service/auth
